policy: add class representation of XML
XML is not the easiest thing to parse (or create). However, SROS 2
policies are nicely grouped into objects that lend themselves well to a
class representation. Add a set of user-friendly classes that represents
an SROS 2 policy and its components that can load and dump themselves
from and to XML. Also replace important strings with enums such that
they cannot be mistaken. Update `generate_policy` to use this new
system. Finally, cover the whole thing in tests.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>