tests: use older pytest
ROS 2 tests are limited by the version of packages available in Ubuntu
Bionic, not the ones available in pip. As a result, we need to make sure
we only take advantage of the features as of pytest v3.3.2. A side
effect of this is that we must use `tmpdir_factory` as opposed to
`tmp_path_factory`, which wasn't introduced until v4.3.0.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>