Remove the use of pkg_resources.
Importing pkg_resources is slow; really slow.  pkg_resources
is also somewhat deprecated.  For both of these reaons, we
would prefer to move away from pkg_resources and start using
importlib instead.

sros2 was using 'resource_filename' from pkg_resources.  The
equivalent in importlib is importlib.resources, but there is a catch.
As of Python 3.8, importlib.resources does not support getting
data from subdirectories.  The workaround is to make each
level of the directory structure a python module, so that the
module name can be looked up.  Python 3.9 will have support for
directories for resources, but that won't help on Ubuntu Focal.
More information is in https://gitlab.com/python-devs/importlib_resources/issues/58

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>