Use S/MIME API from cryptography instead of SSL bindings
The SROS2 utilities use the Cryptography library, which doesn't
officially support S/MIME. As a result, we had to hit the SSL bindings
directly to get that functionality. These bindings were removed in v3.1
of Cryptography, which means the SROS2 utilities were stuck on v3.0 at
the latest. However, Cryptography has added an official S/MIME API which
is far easier to use than the raw SSL API. Update the SROS2 utilities to
use it.

Resolve #239
Update ros2/ci#507

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>