Use modern PKCS7 to sign the certificate bytes. (#290)
* Use modern PKCS7 to sign the certificate bytes.

Using the SSLBinding leads to a warning in newer versions
of pycryptography.  Luckily, there is a supported API
called pkcs7 that allows us to do the same thing.  Even
better, this API is supported since pycryptography 3.2,
so this should work on both Ubuntu 22.04 and Ubuntu 24.04
without warnings.

* Keep fallback path for using SSLBinding.

Windows is still using pycryptography 2.9, which doesn't
support the new pkcs7 API.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>