Fix dangerous use of default arguments in python function call.
You cannot safely use = [] as a default argument in
Python; subsequent calls will reuse that object.  It probably
doesn't matter in this case, as these are all "read-only"
arguments, but it is better not to do it at all.  Instead,
we recognize that this function is always called with
all of its arguments, so there is no need for default arguments
at all.

While in here, I noticed that, generate_artifacts was returning
a boolean on an error path, which is not allowed by its signature.  So fix
that as well.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>